fix: MouseParallax, clean code
