+++ playground/components.d.ts
@@ -14,17 +14,10 @@ declare module 'vue' {
     LeviosoDemo: typeof import('./src/components/LeviosoDemo.vue')['default']
     MapControlsDemo: typeof import('./src/components/MapControlsDemo.vue')['default']
     ModelsDemo: typeof import('./src/components/ModelsDemo.vue')['default']
-    OrbitControlsDemo: typeof import('./src/components/OrbitControlsDemo.vue')['default']
-    PointerLockControlsDemo: typeof import('./src/components/PointerLockControlsDemo.vue')['default']
-    PrecipitationDemo: typeof import('./src/components/PrecipitationDemo.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     ScrollDemo: typeof import('./src/components/ScrollDemo.vue')['default']
     TheGizmos: typeof import('./src/components/TheGizmos.vue')['default']
-    TheSmoke: typeof import('./src/components/TheSmoke.vue')['default']
-    TheStars: typeof import('./src/components/TheStars.vue')['default']
-    TheText3D: typeof import('./src/components/TheText3D.vue')['default']
-    TransformControlsDemo: typeof import('./src/components/TransformControlsDemo.vue')['default']
     WobbleMaterial: typeof import('./src/components/WobbleMaterial.vue')['default']
   }
 }
