+++ playground/src/pages/StarsDemo.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { shallowRef } from 'vue'
 import { TresCanvas } from '@tresjs/core'
-import { OrbitControls, Stars } from '@tresjs/cientos'
+import { OrbitControls, Stars, MouseParallax } from '@tresjs/cientos'
 import { SRGBColorSpace, NoToneMapping } from 'three'
 
 const gl = {
@@ -16,6 +16,7 @@ const star = shallowRef<Stars>(null)
 <template>
   <TresCanvas v-bind="gl" ref="canvas">
     <TresPerspectiveCamera :position="[0, 2, 5]" />
+    <MouseParallax />
     <Stars ref="star" :radius="1" />
     <TresGridHelper :args="[10, 10]" />
     <OrbitControls />
