+++ src/core/abstractions/index.ts
@@ -1,20 +1,20 @@
 import Text3D from './Text3D.vue'
 import { useAnimations } from './useAnimations'
 import { Environment } from './useEnvironment/component'
-import { MouseParallax } from './useParallax/component'
+// import { MouseParallax } from './mouseParallax'
 import Stars from './Stars.vue'
 import Precipitation from './Precipitation.vue'
 import Smoke from './Smoke.vue'
 import Levioso from './Levioso.vue'
 import ContactShadows from './ContactShadows.vue'
 
-export * from './useParallax'
+export * from './mouseParallax'
 export * from './useEnvironment'
 export {
   Text3D,
   useAnimations,
   Environment,
-  MouseParallax,
+  // MouseParallax,
   Stars,
   Smoke,
   Levioso,
